Adds visual marker at start/end of the selection and selectionStyle option to Terminal
We introduce a new configuration option - selectionStyle - used to style the text selection layer.
It has three 'modes': plain (the current default), 'mark-end' (add a marked to the end) and 'mark-start' (add a marker to the beginning).

This is useful when integrating with applications that need to allow manipulating either end of the selection range via keyboard