Prevent stack overflow by firing writeSync while writing
Calling writeSync from the onData API event causes a stack overflow exception to
occur because _bufferOffset only gets incremented after the write callback is
fired. One approach to fix this is to just ensure _bufferOffset is incremented
beforehand, but I think it's still generally a bad idea to allow multiple writes
at the same time. So instead we just block them and return immediately if a
write is already in progress, regardless of whether it was sync or not.

Fixes #3265