Experimental re-design of BufferLine.
The new data structure is more flexible and more compact.
However, mapping column number to cell information is no longer O(1).
So we use a cursor as much as possible to reduce the need for that.

This is an early check-in to have a backup and baseline.
A few basic things work but it is NOT READY FOR USE.