Change CircularList implementation to a simple array plus start offset.
This simplifies the implementation, and makes it easier to use
standard Array methods, rather than having to call getCyclicIndex always.
This should make most operations faster. The exception is when we
actually trim at the front, but we "amortize" that.

A benefit is we can defer trimming which allows some flexibility.

We did lose the recycle method.

The name "CircularList" is no longer appropriate.